Change access modifiers and add the BaseBuilder class
Changed the access modifiers of the `BaseGenerator` and `EntityGenerator` classes to `internal`.
Added the `BaseBuilder` class to manage `using` directives, namespaces, and class names.
Introduced methods in the `EntityBuilder` class for setting type arguments, identifier values, and extension references.
Added a new `Usings` property in the `ClassToAugment` class and updated the `GetIdValueExpression` method.
Updated the test files to reflect the changes in the generated code.