Added the TestEntity class and updated generators.
A new `TestEntity` class has been introduced in the `TestNamespace` namespace, which includes a `Name` property and methods for creating and updating instances.

Added a `_properties` list in `BaseBuilder.cs` and a `SetProperties` method.

Updated the `BuildSource` method in `EntityBuilder.cs` to dynamically generate code for properties.

Logic for gathering property information from class definitions has been added to `EntityGenerator.cs`.

Updated the tests in `EntityGeneratorTests` to include the new properties and methods in the `TestEntity` class, and added new tests for properties of type `string` and `int`.

Changed the `Create` and `Update` methods to support new parameters and update logic.